msm: IPA: devote IMP on 4G ssr
There is one missing case which on 4G ssr happens, mhi proxy is still
in lpm exit and still hold the IMP clock. When new probe cb happens, it
will get double voted and never go to power collapse. The fix here is
to not move the imp state to PROBE so in mhi remove, we will devote
the clock.

Change-Id: I4dc487d322406626492d3ec09487efad8bd02100
Signed-off-by: Bojun Pan <bojunp@codeaurora.org>